Move copying non-TS files to its own file.
