+++ lib/common/application/buildApplication.ts
@@ -1,4 +1,5 @@
 import { compileWithTypeScript } from './compileWithTypeScript';
+import { copyNonTypeScriptFiles } from './copyNonTypeScriptFiles';
 import { isTypeScript } from 'is-typescript';
 import path from 'path';
 import { cp, mkdir, rm } from 'shelljs';
@@ -9,6 +10,7 @@ const buildApplication = async function ({ applicationDirectory, buildDirectoryO
 }): Promise<void> {
   const serverDirectory = path.join(applicationDirectory, 'server');
   const buildDirectory = buildDirectoryOverride ?? path.join(applicationDirectory, 'build');
+  const buildServerDirectory = path.join(buildDirectory, 'server');
 
   rm('-rf', buildDirectory);
 
@@ -18,11 +20,16 @@ const buildApplication = async function ({ applicationDirectory, buildDirectoryO
       targetDirectory: buildDirectory
     });
 
+    await copyNonTypeScriptFiles({
+      sourceDirectory: serverDirectory,
+      targetDirectory: buildServerDirectory
+    });
+
     return;
   }
 
-  mkdir('-p', path.join(buildDirectory, 'server'));
-  cp('-r', `${serverDirectory}/*`, path.join(buildDirectory, 'server'));
+  mkdir('-p', buildServerDirectory);
+  cp('-r', `${serverDirectory}/*`, buildServerDirectory);
 };
 
 export { buildApplication };
